refactor: move error icons into the corresponding components
