--- src/modules/common/components/Icons/ImgError.tsx
@@ -1,47 +1,0 @@
-import { SvgIcon, SvgIconProps } from '@material-ui/core';
-import { makeStyles, Theme } from '@material-ui/core';
-
-const bgColor = '#E0E0E0'
-export const ImgErrorIcon = (props: SvgIconProps) => {
-  return (
-    <SvgIcon {...props} viewBox="0 0 50 70">
-      <rect width="50" height="70" rx="8" fill="white" />
-      <path fillRule="evenodd" clipRule="evenodd" d="M45 12C45 14.2091 43.2091 16 41 16C38.7909 16 37 14.2091 37 12C37 9.79086 38.7909 8 41 8C43.2091 8 45 9.79086 45 12ZM42.9725 35H6.91117C6.10161 35 5.63982 34.0882 6.10486 33.4256L18.9183 15.1664C19.3165 14.599 20.1572 14.599 20.5554 15.1664L27.4268 24.9582C27.7991 25.4887 28.5698 25.5304 28.9971 25.0432L32.62 20.9132C33.035 20.4401 33.7787 20.4634 34.1633 20.9616L43.7564 33.3889C44.2638 34.0463 43.8029 35 42.9725 35ZM7 43C6.44771 43 6 43.4477 6 44C6 44.5523 6.44771 45 7 45H43C43.5523 45 44 44.5523 44 44C44 43.4477 43.5523 43 43 43H7ZM6 51C6 50.4477 6.44771 50 7 50H43C43.5523 50 44 50.4477 44 51C44 51.5523 43.5523 52 43 52H7C6.44771 52 6 51.5523 6 51ZM7 57C6.44771 57 6 57.4477 6 58C6 58.5523 6.44771 59 7 59H29C29.5523 59 30 58.5523 30 58C30 57.4477 29.5523 57 29 57H7Z" fill="#E0E0E0" />
-    </SvgIcon>
-  );
-};
-
-export const VideoErrorIcon = (props: SvgIconProps) => {
-  return (
-    <SvgIcon {...props} viewBox="0 0 50 70">
-      <rect width="50" height="70" rx="8" fill="white" />
-      <path fillRule="evenodd" clipRule="evenodd" d="M25 33C32.1797 33 38 27.1797 38 20C38 12.8203 32.1797 7 25 7C17.8203 7 12 12.8203 12 20C12 27.1797 17.8203 33 25 33ZM7 43C6.44772 43 6 43.4477 6 44C6 44.5523 6.44772 45 7 45H43C43.5523 45 44 44.5523 44 44C44 43.4477 43.5523 43 43 43H7ZM6 51C6 50.4477 6.44772 50 7 50H43C43.5523 50 44 50.4477 44 51C44 51.5523 43.5523 52 43 52H7C6.44772 52 6 51.5523 6 51ZM7 57C6.44772 57 6 57.4477 6 58C6 58.5523 6.44772 59 7 59H29C29.5523 59 30 58.5523 30 58C30 57.4477 29.5523 57 29 57H7Z" fill="#E0E0E0" />
-      <path d="M30.7882 19.6738C31.3704 20.0707 31.3704 20.9293 30.7882 21.3262L22.5633 26.9341C21.8995 27.3867 21 26.9113 21 26.1079L21 14.8921C21 14.0887 21.8995 13.6133 22.5633 14.0659L30.7882 19.6738Z" fill="white" />
-    </SvgIcon>
-  );
-};
-
-export const MusicErrorIcon = (props: SvgIconProps) => {
-  return (
-    <SvgIcon {...props} viewBox="0 0 51 70">
-      <rect x="0.206543" width="50" height="70" rx="8" fill="white" />
-      <path fillRule="evenodd" clipRule="evenodd" d="M19.3564 12.6562L36.2136 8.44189L36.137 28.9131C36.0216 30.6362 33.9894 32.0084 31.4993 32.0084C28.9351 32.0084 26.8564 30.5533 26.8564 28.7584C26.8564 26.9634 28.9351 25.5084 31.4993 25.5084C32.5127 25.5084 33.4502 25.7356 34.2136 26.1213V15.9558L21.3564 19.1701L21.2801 32.6274C21.1646 34.3505 19.1324 35.7227 16.6424 35.7227C14.0782 35.7227 11.9995 34.2676 11.9995 32.4727C11.9995 30.6777 14.0782 29.2227 16.6424 29.2227C17.6556 29.2227 18.5931 29.4499 19.3564 29.8355V12.6562ZM7.20654 43C6.65426 43 6.20654 43.4477 6.20654 44C6.20654 44.5523 6.65426 45 7.20654 45H43.2065C43.7588 45 44.2065 44.5523 44.2065 44C44.2065 43.4477 43.7588 43 43.2065 43H7.20654ZM6.20654 51C6.20654 50.4477 6.65426 50 7.20654 50H43.2065C43.7588 50 44.2065 50.4477 44.2065 51C44.2065 51.5523 43.7588 52 43.2065 52H7.20654C6.65426 52 6.20654 51.5523 6.20654 51ZM7.20654 57C6.65426 57 6.20654 57.4477 6.20654 58C6.20654 58.5523 6.65426 59 7.20654 59H29.2065C29.7588 59 30.2065 58.5523 30.2065 58C30.2065 57.4477 29.7588 57 29.2065 57H7.20654Z" fill={bgColor} />
-    </SvgIcon>
-  );
-};
-
-export const useImgErrorStyles = makeStyles<
-  Theme,
-  { error: boolean }
->(theme => ({
-  root:{
-    backgroundColor: props => props.error ? bgColor : '',
-  },
-  img: {
-    position: 'absolute',
-    top: '50%',
-    left: '50%',
-    fontSize: '50px',
-    transform: 'translate3d(-50%, -50%, 0px)',
-  },
-}));
