+++ src/modules/common/components/VideoPlayer/useVideoPlayerStyles.ts
@@ -14,6 +14,9 @@ export const useVideoPlayerStyles = makeStyles<Theme>(theme => ({
       minHeight: '100%',
     },
   },
+  rootError: {
+    backgroundColor: '#E0E0E0',
+  },
   player: {
     width: '100%',
     height: 'auto',
@@ -26,5 +29,13 @@ export const useVideoPlayerStyles = makeStyles<Theme>(theme => ({
   none: { objectFit: 'none' },
   'scale-down': {
     objectFit: 'scale-down',
+  },
+
+  errorIcon: {
+    position: 'absolute',
+    top: '50%',
+    left: '50%',
+    fontSize: '50px',
+    transform: 'translate3d(-50%, -50%, 0px)',
   },
 }));
